Extract `approve_market` tests
